safari-534.53-branch/LayoutTests/http/tests/security/xss-DENIED-contentWindow-eval-expected.txt

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL about:blank from frame with URL about:blank. Domains, protocols and ports must match.

This test passes if alert() is not called. 
